# apps/backend/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "OMNIVIS API",
    "description": "API para orquestração logística de solidariedade em desastres naturais",
    "version": "1.0.0",
    "contact": {
      "name": "OMNIVIS Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://omnivis-backend.railway.app",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
...
    "/api/v1/voluntarios": {
      "get": {
        "summary": "Listar todos os voluntários",
        "description": "Retorna uma lista de todos os voluntários cadastrados",
        "responses": {
          "200": {
            "description": "Lista de voluntários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Voluntario"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar voluntários"
          }
        }
      },
      "post": {
        "summary": "Criar novo voluntário",
        "description": "Cadastra um novo voluntário no sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoluntarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voluntário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao criar voluntário"
          }
        }
      }
    },
    "/api/v1/voluntarios/{id}": {
      "get": {
        "summary": "Buscar voluntário por ID",
        "description": "Retorna os detalhes de um voluntário específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voluntário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          },
          "404": {
            "description": "Voluntário não encontrado"
          },
          "500": {
            "description": "Erro ao buscar voluntário"
          }
        }
      },
      "put": {
        "summary": "Atualizar voluntário",
        "description": "Atualiza os dados de um voluntário existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoluntarioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voluntário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Voluntario"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao atualizar voluntário"
          }
        }
      },
      "delete": {
        "summary": "Deletar voluntário",
        "description": "Remove um voluntário do sistema",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Voluntário deletado"
          },
          "500": {
            "description": "Erro ao deletar voluntário"
          }
        }
      }
    },
...
    "/api/v1/packs": {
      "get": {
        "summary": "Listar todos os packs",
        "description": "Retorna uma lista de todos os packs cadastrados",
        "responses": {
          "200": {
            "description": "Lista de packs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pack"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar packs"
          }
        }
      },
      "post": {
        "summary": "Criar novo pack",
        "description": "Cadastra um novo pack no sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pack criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao criar pack"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Voluntario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telefone": {
            "type": "string"
          },
          "localizacao": {
            "type": "string",
            "description": "Coordenadas geográficas no formato 'lat,lng'",
            "example": "-23.5505,-46.6333"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "disponibilidades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Disponibilidade"
            }
          }
        }
      },
      "VoluntarioCreate": {
        "type": "object",
        "required": ["nome", "email", "telefone", "localizacao"],
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telefone": {
            "type": "string",
            "minLength": 10
          },
          "localizacao": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d+,-?\\d+\\.\\d+$"
          }
        }
      },
      "VoluntarioUpdate": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telefone": {
            "type": "string",
            "minLength": 10
          },
          "localizacao": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d+,-?\\d+\\.\\d+$"
          }
        }
      },
      "Disponibilidade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": ["Saude", "Logistica", "Engenharia"]
          },
          "ativo": {
            "type": "string",
            "description": "Ativo ou recurso disponível (Barco, Caminhao, Drone, CNH-D, etc.)"
          },
          "voluntarioId": {
            "type": "string"
          },
          "voluntario": {
            "$ref": "#/components/schemas/Voluntario"
          }
        }
      },
...
      "Pack": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": ["Reconstrucao", "Limpeza", "Logistica"]
          },
          "membros": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "IDs dos voluntários"
          },
          "disponivel": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "PackCreate": {
        "type": "object",
        "required": ["nome", "tipo", "membros"],
        "properties": {
          "nome": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": ["Reconstrucao", "Limpeza", "Logistica"]
          },
          "membros": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "disponivel": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
    }
  }
}